Fix ruby-test-find-target-filename bug.
This function could not found correct target file when more than
two candidate exist in some case.

Here is a example:
target-filename-candidates is a list which exist three path strings
("not_existed/path1.rb" "existed/path2.rb" "not_existed/path3")

For previous code base, following code will execute on every element in list.
(if (file-exists-p filename)
  filename
nil)

this will result in the dolist return nil. (because the last path is not exist.)
And, finally result in `(car target-filename-candidates)' will be used, this is not correct.